+++ src/main/java/com/example/PassGuard/controller/SecretController.java
@@ -4,8 +4,8 @@ package com.example.PassGuard.controller;
 import com.example.PassGuard.dto.SecretDto;
 import com.example.PassGuard.model.Secret;
 import com.example.PassGuard.model.User;
-import com.example.PassGuard.service.SecretService;
-import com.example.PassGuard.service.UserService;
+import com.example.PassGuard.service.SecretService.SecretService;
+import com.example.PassGuard.service.UserService.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
