+++ src/main/java/com/example/PassGuard/service/SecretService/SecretService_interface.java
@@ -1,4 +1,4 @@
-package com.example.PassGuard.service.Interface;
+package com.example.PassGuard.service.SecretService;
 
 import com.example.PassGuard.dto.SecretDto;
 import com.example.PassGuard.model.Secret;
