+++ src/main/java/com/example/PassGuard/service/UserService/UserService_Interface.java
@@ -1,4 +1,4 @@
-package com.example.PassGuard.service.Interface;
+package com.example.PassGuard.service.UserService;
 
 import com.example.PassGuard.model.User;
