+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.application.name=PassGuard
 # MySQL Configuration
 spring.datasource.url=jdbc:mysql://localhost:3306/secretmanager?createDatabaseIfNotExist=true
 spring.datasource.username=root
-spring.datasource.password=asm_tamzid
+spring.datasource.password=simec123
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
